feat: complete scheduler

+++ entry.go
@@ -3,5 +3,5 @@ package workersync
 type entry struct {
 	name     string
 	sequence string
-	maxRun   uint
+	maxRun   int64
 }
